拆分 gui_main 岗位配置页构建
